2-bit HW cursor support
Older Tegra generations only support a classic 2-color cursor acceleratable
in HW. A recent kernel with a HW cursor support is required. The HW cursor is
preferred by default now, to maintain compatibility with older kernels we are
checking the kernel support for a HW cursor support early and fallbacking to
the SW cursor.